Add macros
Macros are just like lambdas, but instead of telling how to compute a value, it
tells how to compute another Lisp expression (its expansion) which will in turn
compute the value.

The arguments of macros are not evaluated when calling it, they are passed
literally.

This allows the user to define very cool constructs like `defun' and `defmacro',
instead of being defined as C primitives.

For more information, see the Macros section in the Emacs Lisp manual, which was
used as inspiration:
https://www.gnu.org/software/emacs/manual/html_node/elisp/Macros.html