Fix buffer overflow on read_expr()
If the input is divisible by READ_BUFSZ, it would exactly fit in the allocated
buffer, and no reallocations would be made. The problem is that an extra
null-terminator is written outside of the loop (current line 112). We could add
an extra check there and an extra call to `realloc', or we can just make sure
there is always at least one extra space for the null terminator in the
buffer. I think the second approach is better.