Add expr_list_free() function
The previous version of expr_free() dealt with adjacent Expr structures, and
that wasn't obvious or logical, specially when looking at calls.

This new version is a wrapper for expr_free() and is used in the few places when
an actual linked list of Expr structures needs to be freed. Now there is no
ambiguity about whether we are freeing a single Expr structure or a linked list.

I have checked all calls to expr_free() and I have tested this commit with
valgrind, but I might have missed some edge case.