Remove ENV_FLAG_INVALID from 'EEnvBindingFlags' enum
Return ENV_FLAG_NONE from 'env_get_flags'. We still should return a unique value
to let the caller know that the symbol is not in the environment (rather than
saying that this symbol has no flags set).

Since it was -1 (i.e. 0xFFFFFFFF) it was the same as having all of the flags
set. This did not cause problems because the call to 'env_get_flags' came from
'eval_function_call' (through 'is_special_form'), and it checks whether or not
the string evaluates to a valid function symbol or not. In other words, all
unbound symbols were being marked as "special forms", although this currently
didn't cause any problems.