Fix bug when reading isolated quoted expressions
By "isolated quoted expressions" I mean, for example:

    '(foo bar)
    `(foo bar)
    ,(foo)      // Invalid syntax, but assertion shouldn't fail.

This happened because we were interpreting quotes as normal symbol characters,
and since '(' is a token separator, it stopped early. The lexer then generated
an array like:

    [ QUOTE, EOF ]

It was then passed to the parser, and an assertion in 'wrap_in_call' failed.

Note that this bug didn't affect the following expressions, since it read all
characters as a single symbol (which was then parsed correctly):

    'foo
    `foo
    ,foo