Free all expression members in 'pool_close'
The values of the expressions might be pointers to allocated data (e.g. with
strings or lambdas), so we have to free that too. Since the expression "owns" a
unique pointer to its allocated data (i.e. the another expression won't point to
the same allocated string; perhaps to a copy) we should be able to free it
safely.