Remove unnecessary calls to 'expr_free'
This commit is not just for proving that garbage collection works. In future
commits, many calls to 'expr_clone' will be removed, and most expressions will
be re-used or modified directly (depending on the context, obviously). I decided
to keep these two commits (removing frees and removing clones) separate for
readability.