Only free heap members in 'pool_close' and 'pool_free'
This, along with the previous commit (9a17f00), fixes a bug where 'pool_close'
tried to free expression members multiple times. The problem was caused in the
following case:

1. There was a symbol (for example) in the pool, whose string was freed from the
   heap when 'pool_close' was called.
2. After freeing that symbol, another expression of type 'EXPR_PARENT' was freed
   by 'pool_close'. However, this list contained a pointer to the symbol
   expression whose string we just freed from 'pool_close'. The old
   'free_expr_members' tried freeing the expression with 'expr_list_free', which
   called 'pool_free', which then freed the string again.

This is normally (i.e. when garbage-collecting) not a problem because the string
of the first symbol would be freed through 'pool_free', which would set
'NODE_FLAG_FREE' in the symbol's 'PoolNode'. Therefore, whenever the list is
reached, even if 'expr_list_free' tried freeing each element, 'pool_free' would
notice that it's already been freed (since it checks the flag).

The problem is that this flag is never set whenever we freed the members of the
expression from 'pool_close' (calling 'free_expr_members', not through
'pool_free'), so 'NODE_FLAG_FREE' was never set.

There are two possible solutions:

1. Simply mark the expression whose members we are freeing from 'pool_close'
   with 'NODE_FLAG_FREE'.
2. Don't call 'expr_list_free' from 'free_expr_members' entirely.

Although only one of them is necessary, I chose to add both, since it was not a
good idea to free lists recursively here anyway; specially since we want to
start reusing pointers, and we were not checking if they were in use somewhere
else (we are not checking for 'GCMARKED', but even if it was set, maybe we
weren't called from the garbage collector).

Changes:

* Rename 'free_expr_members' to 'free_heap_expr_members'.
* Don't free expressions recursively from 'free_heap_expr_members'.
* Mark expressions with freed members with 'NODE_FLAG_FREE' from 'pool_close'.