Don't mark 'g_expr_pool' or 'array_start' as 'NOACCESS'
The point of using valgrind is avoiding memory leaks and invalid memory accesses
in the pool (e.g. overflowing the size of an 'Expr', double-frees, etc.), not
controlling when we can access the 'ExprPool' or 'ArrayStart' structures. It's
very hard to access these structures by mistake; if we access them, for example
from the garbage collector, it's because we need to. However, the 'flags' member
of the 'PoolNode' structures, can be easily overwritten by casting a 'Expr*',
for example.

This commit doesn't fix anything, it just removes unnecessary macro calls.

* Remove 'POOL_FOREACH_ARRAYSTART' and 'POOL_FOREACH_ARRAYSTART_END' macros,
  replace with normal for-loops.
* Remove unnecessary calls to 'VALGRIND_MAKE_MEM_DEFINED' and
  'VALGRIND_MAKE_MEM_NOACCESS' from 'expr_pool.c'.