Mark globals from 'env.h' in 'repl_until_eof'
Since we store copies in the environment and not the actual pointers in the
globals, the garbage collector was freeing the globals, and they were being
overwritten with other values. When functions like 'prim_equal' tried to return
clones of 'g_tru', it returned an invalid value (like 'car', for example).