Don't mark environments "in use" in 'gc_mark_env_and_parents'
This greatly improves performance, since it doesn't have to recursively check
common parents each time. For example, consider the following environment
hierarchy:

    (global-env)
       |
       |- (lambda-env-0)
       |
       `- (lambda-env-1)

Before this commit, when marking 'lambda-env-0', we would mark all parents, in
this case 'global-env'; and when marking 'lambda-env-1', we would also try to
mark 'global-env', which would make many calls to 'gc_mark_expr' that would then
return immediately, since the expressions were already marked as "free". These
function calls and conditionals lower the performance of the garbage collector.

After this commit, however we would mark 'lambda-env-0' and its 'global-env'
parent, and when marking 'lambda-env-1', we would skip over the parent, since
its 'is_used' member was set to true. Naturally, both of these examples assume
that 'global-env' wasn't already marked when 'lambda-env-0' was marked, which is
normally not the case, but could be.

Note that this method assumes that, if an environment is marked as "used", all
of its expressions have been marked with 'gc_mark_expr' too, just like
'gc_mark_env' does (through 'gc_mark_env_contents'). This is currently the case;
we just have to be careful if this is changed in the future. In other words, we
assume that an environment (and its contents) can't be "partially" marked.

To see the actual performance impact on one of the tests, you can use the
samply[1] profiler (in my case 0.13.1) with the following command:

    $ samply record ./sl ./test/arith.lisp

Then compare the amount of call groups with the following form:

    1. 'gc_mark_env_contents'
    2. 'gc_mark_expr'
    3. 'gc_mark_env_and_parents'
    4. 'gc_mark_env'

[1] https://github.com/mstange/samply