Move 'mem_realloc' to source, receive double-pointer
Before this commit, the macro used to overwrite the variable that it
received in-place. After this commit, it receives a pointer to the
pointer that was returned by an allocation function such as 'mem_alloc'
or 'mem_calloc', and modifies its value.

This change might make some calls to 'mem_realloc' a bit more confusing,
specially when it's called from a function that already received a
double-pointer, like 'read_user_string'.